refactor: ActivityPeriodField 분리
