+++ src/components/form/Presentation/index.tsx
@@ -1,4 +1,3 @@
-import BubblePointIcon from '@assets/svg/bubble_point.svg';
 import CheckSelectedIcon from '@assets/svg/checkBox/form_selected.svg';
 import CheckUnselectedIcon from '@assets/svg/checkBox/form_unselected.svg';
 import CancelIcon from '@assets/svg/x.svg';
@@ -11,10 +10,9 @@ import TitleField from '@components/form/Presentation/TitleField';
 import WelcomeMessageField from '@components/form/Presentation/WelcomeMessageField';
 import { colors } from '@sopt-makers/colors';
 import { fontsObject } from '@sopt-makers/fonts';
-import { IconAlertCircle } from '@sopt-makers/icons';
 import { useDialog } from '@sopt-makers/ui';
 import { useRouter } from 'next/router';
-import React, { ChangeEvent, ReactNode, useEffect, useRef, useState } from 'react';
+import React, { ChangeEvent, ReactNode, useRef } from 'react';
 import { FieldError, FieldErrors } from 'react-hook-form';
 import { styled } from 'stitches.config';
 import CalendarInputForm from '../Calendar';
@@ -24,6 +22,7 @@ import HelpMessage from '../HelpMessage';
 import Label from '../Label';
 import Textarea from '../Textarea';
 import TextInput from '../TextInput';
+import ActivityPeriodField from './ActivityPeriodField';
 import CoLeader from './CoLeader';
 import DescriptionField from './DescriptionField';
 interface PresentationProps {
@@ -56,55 +55,12 @@ function Presentation({
 }: PresentationProps) {
   const router = useRouter();
   const { open } = useDialog();
-  const [isSoptScheduleOpen, setIsSoptScheduleOpen] = useState(false);
-  const soptScheduleRef = useRef<HTMLDivElement | null>(null);
   const { handleChangeFile, handleDeleteFile, handleAddFiles } = useImageHandler({
     onChangeImage: handleChangeImage,
     onDeleteImage: handleDeleteImage,
   });
 
   const isEdit = router.asPath.includes('/edit');
-
-  const schedule: React.ReactNode = (
-    <>
-      • 1~8차 세미나 <br />
-      &nbsp;&nbsp;&nbsp;2025.04.05 ~ 2025.06.21 <br />
-      • 솝커톤 <br />
-      &nbsp;&nbsp;&nbsp;2025.05.17 ~ 2025.05.18 <br />
-      • 네트워킹 행사 <br />
-      &nbsp;&nbsp;&nbsp;2025.05.31 <br />
-      • 기획 경선 <br />
-      &nbsp;&nbsp;&nbsp;2025.06.07 <br />
-      • 앱잼 <br />
-      &nbsp;&nbsp;&nbsp;2025.06.14 ~ 2025.07.19 <br />
-      • 종무식 <br />
-      &nbsp;&nbsp;&nbsp;2025.07.26
-    </>
-  );
-
-  const soptScheduledialogOption: DialogOptionType = {
-    title: 'SOPT 공식 일정',
-    description: schedule,
-    type: 'default',
-  };
-
-  const handleSoptScheduleOpen = (isOpen: boolean) => {
-    window.innerWidth <= 768 ? open(soptScheduledialogOption) : setIsSoptScheduleOpen(isOpen);
-  };
-
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (soptScheduleRef.current && !soptScheduleRef.current.contains(event.target as Node)) {
-        setIsSoptScheduleOpen(false);
-      }
-    };
-    document.addEventListener('mousedown', handleClickOutside);
-
-    // 컴포넌트가 언마운트될 때 이벤트 리스너 제거
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [soptScheduleRef]);
 
   const formRef = useRef<HTMLFormElement>(null);
 
@@ -141,91 +97,7 @@ function Presentation({
         <SectionLine />
         <div style={{ display: 'flex', flexDirection: 'column', gap: '52px' }}>
           {/* 활동 정보 - 모임 기간 */}
-          <div>
-            <div style={{ display: 'flex', justifyContent: 'space-between' }}>
-              <Label required={true} size="small">
-                활동 기간
-              </Label>
-              <div
-                ref={soptScheduleRef}
-                style={{ display: 'flex', gap: '4px', position: 'relative' }}
-                onMouseEnter={() => handleSoptScheduleOpen(true)}
-                onMouseLeave={() => handleSoptScheduleOpen(false)}
-              >
-                <div style={{ display: 'flex', gap: '4px', marginRight: '16px' }}>
-                  <SoptNotice>SOPT 공식 일정 확인하기</SoptNotice>
-                  <IconAlertCircle style={{ width: '16px', height: '16px', color: 'gray', cursor: 'pointer' }} />
-                </div>
-                {isSoptScheduleOpen && (
-                  <ToolTipDiv>
-                    <Pointdiv>
-                      <BubblePointIcon />
-                    </Pointdiv>
-                    <TextDiv>
-                      <TextStyle>• 1~8차 세미나: 2025.04.05 ~ 2025.06.21</TextStyle>
-                      <TextStyle>• 솝커톤: 2025.05.17 ~ 2025.05.18</TextStyle>
-                      <TextStyle>• 네트워킹 행사: 2025.05.31</TextStyle>
-                      <TextStyle>• 기획 경선: 2025.06.07</TextStyle>
-                      <TextStyle>• 앱잼: 2025.06.14 ~ 2025.07.19</TextStyle>
-                      <TextStyle>• 종무식: 2025.07.26</TextStyle>
-                    </TextDiv>
-                  </ToolTipDiv>
-                )}
-              </div>
-            </div>
-            <SApplicationFieldWrapper>
-              <SApplicationField>
-                <FormController
-                  name="detail.mDateRange"
-                  render={({ field, formState: { errors } }) => {
-                    const dateError = errors.detail as
-                      | (FieldError & {
-                          mDateRange?: FieldError[];
-                        })
-                      | undefined;
-                    return (
-                      <CalendarInputForm
-                        selectedDate={field.value}
-                        setSelectedDate={field.onChange}
-                        selectedDateFieldName={field.name}
-                        error={
-                          (dateError?.mDateRange as FieldError[])?.[0]?.message ||
-                          (dateError?.mDateRange as FieldError[])?.[1]?.message
-                        }
-                        dateType="startDate"
-                      />
-                    );
-                  }}
-                ></FormController>
-              </SApplicationField>
-              <span style={{ marginTop: '14px' }}>-</span>
-              <SApplicationField>
-                <FormController
-                  name="detail.mDateRange"
-                  render={({ field, formState: { errors } }) => {
-                    const dateError = errors.detail as
-                      | (FieldError & {
-                          mDateRange?: FieldError[];
-                        })
-                      | undefined;
-
-                    return (
-                      <CalendarInputForm
-                        selectedDate={field.value}
-                        setSelectedDate={field.onChange}
-                        dateType="endDate"
-                        error={
-                          (dateError?.mDateRange as FieldError[])?.[0]?.message ||
-                          (dateError?.mDateRange as FieldError[])?.[1]?.message
-                        }
-                        selectedDateFieldName={field.name}
-                      />
-                    );
-                  }}
-                ></FormController>
-              </SApplicationField>
-            </SApplicationFieldWrapper>
-          </div>
+          <ActivityPeriodField />
           {/* 모임 정보 - 진행 방식 소개 */}
           <div>
             <Label required={true} size="small">
@@ -471,18 +343,6 @@ const SForm = styled('form', {
     gap: '56px',
   },
 });
-const STitleField = styled('div', {
-  width: '100%',
-});
-const SFileInputWrapper = styled('div', {
-  display: 'grid',
-  gridTemplateColumns: 'repeat(3, 1fr)',
-  gap: '16px',
-
-  '@media (max-width: 768px)': {
-    gridTemplateColumns: 'repeat(2, 1fr)',
-  },
-});
 const SApplicationFieldWrapper = styled('div', {
   display: 'flex',
   color: '$gray500',
@@ -602,53 +462,4 @@ const SLabelWrapper = styled('div', {
 const SLabelCheckboxWrapper = styled('div', {
   display: 'flex',
   justifyContent: 'space-between',
-});
-
-const SoptNotice = styled('span', {
-  cursor: 'pointer',
-  display: 'inline-block',
-  minWidth: '$125',
-  ...fontsObject.LABEL_4_12_SB,
-  color: '$gray300',
-});
-
-const ToolTipDiv = styled('div', {
-  width: '252px',
-  height: '162px',
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'flex-end',
-
-  position: 'absolute',
-  top: '$20',
-  right: '$0',
-  isolate: 'isolation',
-});
-
-const Pointdiv = styled('div', {
-  display: 'inline-flex',
-  paddingRight: '16px',
-  flexDirection: 'column',
-  alignItems: 'flex-start',
-});
-
-const TextDiv = styled('div', {
-  display: 'inline-flex',
-  padding: '16px',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'flex-start',
-  gap: '4px',
-
-  width: '255px',
-
-  borderRadius: '10px',
-  backgroundColor: '$gray600',
-
-  color: '$gray50',
-});
-
-const TextStyle = styled('p', {
-  ...fontsObject.LABEL_4_12_SB,
-  letterSpacing: '-0.24px',
 });
